requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
pip==21.3.1
setuptools==59.0.1
